Optimize code in login/page.tsx, separate files for better management and easier debugging
